requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
astroid==2.11.7
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.5.1
Django==4.1
django-admin-searchable-dropdown==1.2
django-environ==0.9.0
django-humanize==0.1.2
django-nested-admin==3.4.0
django-tz-detect==0.4.0
djangorestframework==3.14.0
executing==0.10.0
frozenlist==1.3.1
github==1.2.7
humanize==4.3.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
matplotlib-inline==0.1.3
mccabe==0.7.0
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.55
Pygments==2.12.0
PyJWT==2.5.0
pylint==2.14.5
PyNaCl==1.5.0
pyOpenSSL==22.0.0
python-monkey-business==1.0.0
pytz==2022.2.1
redgreenunittest==0.1.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
tomli==2.0.1
tomlkit==0.11.1
traitlets==5.3.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tz==0.2.2
urllib3==1.26.12
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
zope.interface==5.4.0
